refactor(optims): group all optimization related stuff in source/optims
