+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "publiopti",
+  "name": "@publicodes/tools",
   "version": "0.1.28",
-  "description": "An optimizer lib for Publicodes models",
+  "description": "A set of utility functions to build tools around Publicodes models",
   "type": "module",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -11,33 +11,40 @@
       "import": "./dist/index.js",
       "require": "./dist/index.cjs",
       "types": "./dist/index.d.ts"
+    },
+    "./optims": {
+      "import": "./dist/optims/index.js",
+      "require": "./dist/optims/index.cjs",
+      "types": "./dist/optims/index.d.ts"
     }
   },
   "files": [
     "dist"
   ],
-  "repository": "git@github.com:datagir/publiopti.git",
-  "bugs": "https://github.com/datagir/publiopti/issues",
+  "repository": "git@github.com:incubateur-ademe/publicodes-tools.git",
+  "bugs": "https://github.com/incubateur-ademe/publicodes-tools/issues",
   "keywords": [
-    "publicodes",
     "optimization",
+    "publicodes",
+    "tooling",
+    "scripting",
     "constant folding"
   ],
   "author": "<emile.rolley@tuta.io>",
   "license": "MIT",
   "scripts": {
     "build": "tsup",
-    "build:watch": "tsup --watch",
+    "watch": "tsup --watch",
     "clean": "rm -rf dist",
     "test": "jest"
   },
   "dependencies": {
-    "@types/node": "^18.11.18"
+    "@types/node": "^18.11.18",
+    "publicodes": "^1.0.0-beta.70"
   },
   "devDependencies": {
     "@types/jest": "^29.2.5",
     "jest": "^29.4.1",
-    "publicodes": "^1.0.0-beta.66",
     "ts-jest": "^29.0.4",
     "ts-node": "^10.9.1",
     "tsup": "^6.5.0",
@@ -45,7 +52,8 @@
   },
   "tsup": {
     "entry": [
-      "source/index.ts"
+      "source/index.ts",
+      "source/optims/index.ts"
     ],
     "format": [
       "cjs",
@@ -53,7 +61,9 @@
     ],
     "sourceMap": true,
     "dts": true,
-    "clean": true,
-    "globalName": "publiopti"
+    "clean": true
+  },
+  "publishConfig": {
+    "access": "public"
   }
 }
