+++ source/index.ts
@@ -1,2 +1,1 @@
-export { constantFolding } from "./constantFolding";
 export { disabledLogger, getRawNodes } from "./commons";
