+++ source/optims/index.ts
@@ -1,0 +1,1 @@
+export { constantFolding } from "./constantFolding";
